Updated: Core, Unit Tests, Framework
Moved all Ready and Wait Commands on to InterfaceControl to allow Dialog Box support.

Added Support for Keypress on button in Dialog Box, Click does work but only if the button is visible, both methods remain. To press the space key on a button use PressButton or ClickButton to use mouse. Using the mouse method was causing unit test failures.

Added WaitForDialog on ApplicationUnderTest class, returns as MessageBoxDialog class the caption, text, number of button and a PressButton method.